pyproject.toml
[project]
name = "arc-raiders-helper"
version = "0.1.0"
description = "Item recommendation overlay for Arc Raiders"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
arc-helper = "arc_helper.main:main"
arc-calibrate = "arc_helper.calibrate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arc_helper"]

[tool.ruff]
line-length = 88 # as black
indent-width = 4
target-version = "py311"
exclude = [
 "modules/mci/relo.py",
 "modules/mci/relo_v2.py",
 "modules/helpers/helpers.py",
 "modules/hadoop/best_servers_svk_mapping.py",
]

[tool.ruff.format]
docstring-code-format = true

# https://docs.astral.sh/ruff/rules/
[tool.ruff.lint]
preview = true
select = [
 # "AIR",   # Airflow (AIR)
 # "ERA",   # eradicate (ERA)
 "FAST", # FastAPI (FAST)
 "YTT",  # flake8-2020 (YTT)
 # "ANN",   # flake8-annotations (ANN)
 "ASYNC", # flake8-async (ASYNC)
 # "S",     # flake8-bandit (S)
 "BLE", # flake8-blind-except (BLE)
 "FBT", # flake8-boolean-trap (FBT)
 "B",   # flake8-bugbear (B)
 "A",   # flake8-builtins (A)
 "COM", # flake8-commas (COM)
 "C4",  # flake8-comprehensions (C4)
 # "CPY",   # flake8-copyright (CPY)
 "DTZ", # flake8-datetimez (DTZ)
 # "T10",   # flake8-debugger (T10)
 # "DJ",    # flake8-django (DJ)
 "EM", # flake8-errmsg (EM)
 # "EXE",   # flake8-executable (EXE)
 # "FIX",   # flake8-fixme (FIX)
 "FA",  # flake8-future-annotations (FA)
 "INT", # flake8-gettext (INT)
 "ISC", # flake8-implicit-str-concat (ISC)
 "ICN", # flake8-import-conventions (ICN)
 "LOG", # flake8-logging (LOG)
 "G",   # flake8-logging-format (G)
 "INP", # flake8-no-pep420 (INP)
 "PIE", # flake8-pie (PIE)
 # "T20",   # flake8-print (T20)
 "PYI",  # flake8-pyi (PYI)
 "PT",   # flake8-pytest-style (PT)
 "Q",    # flake8-quotes (Q)
 "RSE",  # flake8-raise (RSE)
 "RET",  # flake8-return (RET)
 "SLF",  # flake8-self (SLF)
 "SIM",  # flake8-simplify (SIM)
 "SLOT", # flake8-slots (SLOT)
 "TID",  # flake8-tidy-imports (TID)
 # "TD",    # flake8-todos (TD)
 "TC",   # flake8-type-checking (TC)
 "ARG",  # flake8-unused-arguments (ARG)
 "PTH",  # flake8-use-pathlib (PTH)
 "FLY",  # flynt (FLY)
 "I",    # isort (I)
 "C90",  # mccabe (C90)
 "NPY",  # NumPy-specific rules (NPY)
 "PD",   # pandas-vet (PD)
 "N",    # pep8-naming (N)
 "PERF", # Perflint (PERF)
 "E",    # pycodestyle (E, W) Error (E)
 "W",    # pycodestyle (E, W) Warning (W)
 # "DOC",  # pydoclint (DOC)
 # "D",    # pydocstyle (D)
 "F", # Pyflakes (F)
 # "PGH",  # pygrep-hooks (PGH)
 "PL",   # Pylint (PL)
 "PLC",  # Convention (PLC)
 "PLE",  # Error (PLE)
 "PLR",  # Refactor (PLR)
 "PLW",  # Warning (PLW)
 "UP",   # pyupgrade (UP)
 "FURB", # refurb (FURB)
 "RUF",  # Ruff-specific rules (RUF)
 "TRY",  # tryceratops (TRY)
]
ignore = [
 # "B008",    # read the default from a module-level singleton variable
 "C901",    # `get_skeleton` is too complex (14 > 10)
 "COM812",  # Trailing comma missing, warning: The following rule may cause conflicts when used with the formatter
 "E501",    # ignore line-length
 "PD015",   # Use `.merge` method instead of `pd.merge` function. They have equivalent functionality.
 "PLC0415", # `import` should be at the top-level of a file
 "PLR0911", # Too many return statements (8 > 6)
 "PLR0912", # Too many branches (13 > 12)
 "PLR0913", # Too many arguments in function definition (8 > 5)
 "PLR0914", # Too many local variables (20/15)
 "PLR0915", # Too many statements (81 > 50)
 "PLR0917", # Too many positional arguments (8/5)
 "PLR2004", # Magic value used in comparison, consider replacing `150` with a constant variable
 "PLR0904", # Too many public methods (34 > 20)
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.mypy]
python_version = "3.11"
strict = true
